Refactor Timezone clients to comply to the latest standards
