+++ sdk/maps/azure-maps-timezone/azure/maps/timezone/_patch.py
@@ -6,7 +6,12 @@
 
 Follow our quickstart for examples: https://aka.ms/azsdk/python/dpcodegen/python/customize
 """
-from typing import List
+# pylint: disable=unused-import
+from typing import Union, Any, List
+
+from azure.core.credentials import AzureKeyCredential, TokenCredential
+from azure.core.pipeline.policies import AzureKeyCredentialPolicy
+from azure.maps.timezone import TimezoneClient
 
 __all__: List[str] = []  # Add all objects you want publicly available to users at this package level
 
@@ -18,3 +23,36 @@ def patch_sdk():
     you can't accomplish using the techniques described in
     https://aka.ms/azsdk/python/dpcodegen/python/customize
     """
+
+
+# To check the credential is either AzureKeyCredential or TokenCredential
+def _authentication_policy(credential):
+    authentication_policy = None
+    if credential is None:
+        raise ValueError("Parameter 'credential' must not be None.")
+    if isinstance(credential, AzureKeyCredential):
+        authentication_policy = AzureKeyCredentialPolicy(
+            name="subscription-key", credential=credential
+        )
+    elif credential is not None and not hasattr(credential, "get_token"):
+        raise TypeError(
+            "Unsupported credential: {}. Use an instance of AzureKeyCredential "
+            "or a token credential from azure.identity".format(type(credential))
+        )
+    return authentication_policy
+
+
+class AzureTimezoneClient(TimezoneClient):
+    def __init__(
+        self,
+        credential: Union[AzureKeyCredential, TokenCredential],
+        **kwargs: Any
+    ) -> None:
+
+        super().__init__(
+            credential=credential,  # type: ignore
+            endpoint=kwargs.pop("endpoint", "https://atlas.microsoft.com"),
+            client_id=kwargs.pop("client_id", None),
+            authentication_policy=kwargs.pop("authentication_policy", _authentication_policy(credential)),
+            **kwargs
+        )
